+++ ChangeLog.md
@@ -1,7 +1,7 @@
 2018-08-07
 
 - require asset classes for taskdef registration
-- register executor, asset, taskdef
+- register executor, asset, taskdef, parameter
 
 2018-08-06